trabalho - Copia/requirements.txt
txt
pandas
sqlalchemy
psycopg2-binary
panel
python-dotenv